requirements.txt
google-generativeai
langchain-google-genai
langchain-community
faiss-cpu
pandas
openpyxl